ISSUE #5266 - move and rename selector for getting all values for a oneOf/manyOf
