+++ frontend/src/v5/ui/components/viewer/cards/cardFilters/filterForm/filterFormValues/filterFormValues.component.tsx
@@ -23,7 +23,7 @@ import { ArrayFieldContainer } from '@controls/inputs/arrayFieldContainer/arrayF
 import { useEffect } from 'react';
 import { isArray, isEmpty } from 'lodash';
 import { CardFilterType } from '../../cardFilters.types';
-import { TicketsHooksSelectors } from '@/v5/services/selectorsHooks';
+import { TicketsCardHooksSelectors } from '@/v5/services/selectorsHooks';
 import { useParams } from 'react-router-dom';
 import { ViewerParams } from '@/v5/ui/routes/routes.constants';
 import { MultiSelectMenuItem } from '@controls/inputs/multiSelect/multiSelectMenuItem/multiSelectMenuItem.component';
@@ -61,7 +61,7 @@ export const FilterFormValues = ({ module, property, type }: FilterFolrmValuesTy
 		return '';
 	};
 	const emptyValue = { value: getEmptyValue() };
-	const selectOptions = TicketsHooksSelectors.selectAllValuesByModuleAndProperty(containerOrFederation, module, property, type);
+	const selectOptions = TicketsCardHooksSelectors.selectPropertyOptions(containerOrFederation, module, property, type);
 
 	useEffect(() => {
 		if (!fields.length && maxFields > 0) {
